+++ memsource/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'Gengo'
-__version__ = '0.4.8'
+__version__ = '0.4.9'
 __license__ = 'MIT'