+++ test/test_memsource.py
@@ -9,9 +9,10 @@ class TestMemsource(unittest.TestCase):
     def setUp(self):
         self.url_base = 'https://cloud.memsource.com/web/api/v3/auth/login'
 
-    def check_token(self, m, token):
+    def check_token_and_headers(self, m, token=None, headers=None):
         for api in ('client', 'domain', 'project', 'job', ):
             self.assertEqual(getattr(m, api).token, token)
+            self.assertEqual(getattr(m, api).headers, headers)
 
     @patch.object(requests.Session, 'request')
     def test_init_with_user_name_and_password(self, mock_request):
@@ -25,7 +26,7 @@ class TestMemsource(unittest.TestCase):
 
         username = 'test_username'
         password = 'test_password'
-        self.check_token(Memsource(user_name=username, password=password), token)
+        self.check_token_and_headers(Memsource(user_name=username, password=password), token)
 
         mock_request.assert_called_with(
             constants.HttpMethod.post.value,
@@ -41,7 +42,7 @@ class TestMemsource(unittest.TestCase):
     @patch.object(requests.Session, 'request')
     def test_init_with_token(self, mock_request):
         token = 'test_token'
-        self.check_token(Memsource(token=token), token)
+        self.check_token_and_headers(Memsource(token=token), token)
 
         # When token is given as parameter, never send http request.
         self.assertFalse(mock_request.called)
@@ -50,7 +51,7 @@ class TestMemsource(unittest.TestCase):
     def test_init_with_user_name_and_password_and_token(self, mock_request):
         token = 'test_token'
         m = Memsource(user_name='test user name', password='test password', token=token)
-        self.check_token(m, token)
+        self.check_token_and_headers(m, token)
 
         # When token is given as parameter, never send http request.
         self.assertFalse(mock_request.called)
@@ -58,7 +59,7 @@ class TestMemsource(unittest.TestCase):
     @patch.object(requests.Session, 'request')
     def test_header_memsource_parameter(self, mock_request):
         mock_request.return_value.status_code = 200
-        token = 'test_token'
+        token = None
         mock_request().json.return_value = {
             'token': token,
             'user': {},
@@ -66,13 +67,5 @@ class TestMemsource(unittest.TestCase):
         headers = {
             'Authorization': 'Bearer test_token'
         }
-
-        self.check_token(Memsource(headers=headers), token)
-
-        mock_request.assert_called_with(
-            constants.HttpMethod.post.value,
-            self.url_base,
-            data={'password': None, 'userName': None, 'token': None},
-            timeout=constants.Base.timeout.value,
-            headers={'Authorization': 'Bearer test_token'}
-        )
+        self.check_token_and_headers(Memsource(headers=headers), token, headers)
+        mock_request.called