+++ CHANGELOG.rst
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file. Dates are i
 Unreleased
 ==========
 
+[0.4.9] - 2019-08-23
+====================
+
+Refactored
+-----
+- Refactored `headers` parameter on initiating the Memsource class.
+- Removed `inflection` package.
+
 [0.4.8] - 2019-08-21
 ====================